+++ src/main/java/me/mni_js/springboot_blog/controller/ExampleController.java
@@ -15,7 +15,7 @@ public class ExampleController {
     @GetMapping("/thymeleaf/example")
     public String thymeleafExample(Model model) {
         Person examplePerson = new Person();
-        examplePerson.setId(1);
+        examplePerson.setId(1L);
         examplePerson.setName("홍길동");
         examplePerson.setAge(11);
         examplePerson.setHobbies(List.of("운동", "독서"));
@@ -29,7 +29,7 @@ public class ExampleController {
     @Setter
     @Getter
     class Person {
-        private Integer id;
+        private Long id;
         private String name;
         private int age;
         private List<String> hobbies;
